Add PointNet2 + PointTransformer regression models; flatten repo
- Add selectable PointNet2 and PointTransformer per-point regression
  models via --model {pointnet2,pointtransformer}
- Re-inject per-sample scale feature so the new models can scale
  shrinkage prediction with absolute piece size
- Fix shrinkage target truncation (int64 -> float32) in data loader
- Remove unused cls/partseg/semseg training scripts
- Add conda environment.yml
- Flatten dgcnn/ into repo root to match README layout

+++ data.py
@@ -322,7 +322,7 @@ def load_data_partseg_from_txt(partition, num_points=30000):
             assert pc_data.shape[1] == 4, f"{txt_path} must have 4 columns (x y z label)"
 
             coords_rgb = pc_data[:, 0:3]  # x, y, z
-            seg_labels = pc_data[:, 3].astype(np.int64)  # segmentation labels
+            seg_labels = pc_data[:, 3].astype(np.float32)  # regression target (shrinkage)
         elif pc_data.shape[1] == 3:
             assert pc_data.shape[1] == 3, f"{txt_path} must have 3 columns (x y z)"
